refactor: move internal/script under internal/app
