+++ cmd/script/info.go
@@ -2,10 +2,10 @@ package scriptcmd
 
 import (
 	"fmt"
+	"sfDBTools/internal/app/script"
 	"sfDBTools/internal/cli/deps"
 	"sfDBTools/internal/cli/flags"
 	"sfDBTools/internal/cli/parsing"
-	"sfDBTools/internal/script"
 	"strings"
 
 	"github.com/spf13/cobra"
