+++ cmd/script/run.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"sfDBTools/internal/app/script"
 	"sfDBTools/internal/cli/deps"
 	"sfDBTools/internal/cli/flags"
 	"sfDBTools/internal/cli/parsing"
-	"sfDBTools/internal/script"
 	"sfDBTools/pkg/input"
 	"sort"
 	"strings"
