+++ internal/app/script/extract.go
