+++ internal/app/script/info.go
